% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
benchmark(
  x,
  bench,
  share,
  method = "raking",
  H = NULL,
  time = NULL,
  areas = NULL
)
}
\arguments{
\item{x}{Object of class 'summary.saefit'.}
}
\value{
'benchmark' object.
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
