% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.summary.sae.R
\name{density.summary.saefit}
\alias{density.summary.saefit}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
\method{density}{summary.saefit}(x, grid = TRUE)
}
\arguments{
\item{x}{Object of class 'summary.fitsae'.}

\item{grid, }{logical, indicating whether displaying plots in a grid (TRUE) or in sequence (FALSE).}
}
\value{
Two ggplot2 objects stored in a grid or in sequence.
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
