% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emilia}
\alias{emilia}
\title{Poverty in Emilia-Romagna (Italy) Health Districts}
\format{
Dataframe with 190 observations and 8 variables.

\describe{
\item{\code{id}}{Character, name of the health district.}
\item{\code{prov}}{Character, name of NUTS-3 region related to the district.}
\item{\code{year}}{Numeric, year of the observation.}
\item{\code{hcr}}{Numeric, head-count ratio estimate (used as response variable).}
\item{\code{vars}}{Numeric, sampling variance of head-count ratio estimator.}
\item{\code{n}}{Numeric, area sample size.}
\item{\code{x}}{Numeric, fake covariate.}
\item{\code{pop}}{Numeric, population size of the area.}
}
}
\usage{
emilia
}
\description{
The \code{emilia} dataset consists of a panel on poverty mapping concerning 38 health districts within the Emilia-Romagna region, located in North-East of Italy, with annual observations recorded from 2014 to 2018.
}
\details{
It has been built starting from model-based estimates and related CV freely available on Emilia-Romagna region \href{https://statistica.regione.emilia-romagna.it/documentazione/pubblicazioni/documenti_catalogati/stima-poverta-2009-2018-distretti-sociosanitari-province-emilia-romagna}{website}. Since it is used for illustrative purposes only, such estimates are assumed to be unreliable direct estimates, requiring a SAE procedure.
}
\examples{
\donttest{library(tipsae)
data("emilia")}


}
\keyword{datasets}
