% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emilia_shp}
\alias{emilia_shp}
\title{Shapefile of Emilia-Romagna (Italy) Health Districts}
\format{
A shapefile of class \code{SpatialPolygonsDataFrame}.

\describe{
\item{\code{COD_DIS_SA}}{Code of the health district.}
\item{\code{NAME_DISTRICT}}{Name of the health district. It can be linked to the variable \code{id} in \code{\link{emilia}} and \code{\link{emilia_cs}}}
}
}
\usage{
emilia_shp
}
\description{
The \code{emilia_shp} shapefile consists of a \code{SpatialPolygonsDataFrame} object of 38 health districts within the Emilia-Romagna region, located in the North-East of Italy.
}
\examples{
\donttest{library(tipsae)
data("emilia_shp")}

}
\seealso{
\code{\link{emilia}} and \code{\link{emilia_cs}} for the provided datasets.
}
\keyword{datasets}
