% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_estimates.R
\name{export_estimates}
\alias{export_estimates}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
export_estimates(x, file, digits = 2, type = "all")
}
\arguments{
\item{x}{Object of class 'estimates.saefit' or 'benchmark.saefit'.}
}
\value{

}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
