% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.summary.sae.R
\name{map_sae}
\alias{map_sae}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
map_sae(
  x,
  spatial_df,
  spatial_id_domains,
  match_names = NULL,
  color_palette = c("snow2", "deepskyblue4"),
  quantity = "HB_est",
  time = NULL
)
}
\arguments{
\item{x}{Object of class 'summary.fitsae'.}

\item{map_obj}{sp spatial polygons object.}

\item{id}{id vector ordererd as 'saefit' areas.}

\item{map_dom_id}{denoting name of domains ID variable in patial polygons object.}

\item{color}{vector of two color strings denoting extreme bounds of colors scale to be used.}

\item{label_names}{target parameter name to display as legend title, the default one is "Model-based est.".}
}
\value{
map ggplot2 object (geom_polygon) with colors scaled legend.
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
