% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sae.R
\name{plot.summary.saefit}
\alias{plot.summary.saefit}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
\method{plot}{summary.saefit}(x, size = 2.5, alpha = 0.8, n_bins = 15, grid = TRUE, label_names = NULL)
}
\arguments{
\item{x}{Object of class 'summary.fitsae'.}

\item{size, }{denoting size of scatterplots points (geom_point).}

\item{alpha, }{denoting  opacity of scatterplots points (geom_point).}

\item{n_bins, }{denoting number of bins used for histogram.}

\item{grid, }{logical, indicating whether displaying plots in a grid (TRUE) or in sequence (FALSE).}

\item{label_names, }{model name to display in boxplot x-axis label.}

\item{areas_sample_size}{Areas-specific sample sizes.}
}
\value{
Four ggplot2 objects stored in a grid.
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
