% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sae.R
\name{print.fitsae}
\alias{print.fitsae}
\title{Print Method for a \code{fitsae} Object}
\usage{
\method{print}{fitsae}(x, ...)
}
\arguments{
\item{x}{Object of class \code{fitsae}.}

\item{...}{Currently unused.}
}
\value{
Printed information on a \code{fitsae} object.
}
\description{
The generic method \code{print()} allow to explore relevant outputs of the input object
}
