% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fitsae.R
\name{print.summary_fitsae}
\alias{print.summary_fitsae}
\title{Print Method for a \code{summary_fitsae} Object}
\usage{
\method{print}{summary_fitsae}(x, digits = 3L, ...)
}
\arguments{
\item{x}{Object of class \code{summary_fitsae}.}

\item{digits}{Number of digits to display.}

\item{...}{Currently unused.}
}
\value{
Printed information on a \code{summary_fitsae} object.
}
\description{
The generic method \code{print()} allow to explore relevant outputs of the input object
}
