% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_launcher.R
\name{runShiny_tipsae}
\alias{runShiny_tipsae}
\title{Lauch Shiny App to Performs Small Area Estimation}
\usage{
runShiny_tipsae()
}
\value{
No value returned.
}
\description{
The command launches a Shiny application that assists the user from the data loading step to the export of the outputs. See the vignette for further details.
}
\examples{

library(tipsae)

# Starting the Shiny application
if(interactive()){
 runShiny_tipsae()
}



}
