% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
smoothing(
  data,
  direct_estimates,
  area_id = NULL,
  raw_variance = NULL,
  areas_sample_sizes = NULL,
  additional_covariates = NULL,
  method = c("ols", "gls", "kish"),
  var_function = NULL,
  survey_data = NULL,
  survey_area_id = NULL,
  weights = NULL,
  sizes = NULL
)
}
\arguments{
\item{direct_estimates}{Linear regression formula at the linking level.}
}
\value{
An object of class \code{smoothing.fitsae}.
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
