% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sae.R
\name{summary.saefit}
\alias{summary.saefit}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
\method{summary}{saefit}(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), compute_loo = TRUE, ...)
}
\arguments{
\item{x}{Model fitted via 'fit_sae' function, having class 'saefit'.}
}
\value{
A list containing diagnostics objects.
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
