% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sae.R
\name{fit_sae}
\alias{fit_sae}
\title{Bayesian proportions/(0,1)-measures small area model with Stan}
\usage{
fit_sae(
  formula_fixed,
  domains = NULL,
  disp_direct,
  type_disp = "neff",
  domain_size = NULL,
  data,
  likelihood = "beta",
  prior_reff = "normal",
  prior_coeff = "normal",
  p0_HorseShoe = NULL,
  spatial_error = FALSE,
  spatial_df = NULL,
  temporal_error = FALSE,
  temporal_variable = NULL,
  adapt_delta = 0.95,
  max_treedepth = 10,
  init = "0",
  ...
)
}
\arguments{
\item{formula_fixed}{Linear regression formula at the linking level.}

\item{data}{Data frame containing direct estimates and covariates.}

\item{likelihood}{Selection between different likelihood assumptions at sampling level.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{type}{Selection between different parametrizations.}
}
\value{
An object of class \code{fitsae} containing an object of class \code{stanfit}returned by \code{rstan::sampling}
}
\description{
Bayesian proportions/(0,1)-measures small area model with Stan
}
